refactor the dirs structure
