+++ src/containers/search_result.jsx
